refactor: split ai prompts with current contracts
