+++ package.json
@@ -22,7 +22,7 @@
     "test:security": "npm run test:security:sql && npm run test:security:secrets && npm run test:security:routes",
     "qa:stripe:webhook-e2e": "tsx scripts/qa-stripe-webhook-e2e.ts",
     "test": "npm run test:web",
-    "test:web": "tsx --test src/lib/api/internal-worker.test.ts src/lib/auth/signup-flow.test.ts src/lib/auth/otp-lifecycle.test.ts src/app/api/auth/otp.contract.test.ts src/lib/stripe/client.test.ts src/lib/appstore/config.test.ts src/lib/appstore/verify.test.ts src/lib/appstore/notifications.test.ts src/lib/subscription/status.test.ts src/lib/subscription/display.test.ts src/lib/subscription/reconcile-status.test.ts src/lib/subscription/billing-activation.test.ts src/lib/subscription/stripe-webhook-handlers.test.ts src/lib/supabase/session-cache.test.ts src/lib/db/remote-repository.test.ts src/lib/db/hybrid-repository.test.ts src/lib/db/sync-queue.test.ts src/lib/db/debounced-status-write.test.ts src/lib/projects/load-helpers.test.ts src/lib/stats/calendar.test.ts src/lib/ai/extract-circled-words.test.ts src/lib/ai/extract-highlighted-words.test.ts src/lib/ai/extract-eiken-words.test.ts src/lib/ai/extract-wrong-answers.test.ts src/lib/ai/generate-example-sentences.test.ts src/lib/ai/match-passage-words.test.ts src/lib/ai/prompts.translation-context.test.ts src/lib/lexicon/master-first-scan.test.ts src/lib/lexicon/resolver.test.ts src/lib/lexicon/word-resolution-jobs.test.ts src/lib/words/backfill-japanese.test.ts src/lib/schemas/ai-response.test.ts src/lib/scan/mode-provider.test.ts src/lib/scan/job-result-payload.test.ts src/lib/scan/server-cloud-result-payload.test.ts src/lib/scan/quiz-prefill.test.ts src/lib/scan/post-processing.test.ts src/lib/scan/example-generation.test.ts src/lib/scan/image-extraction.test.ts src/lib/scan/server-cloud-persistence.contract.test.ts src/lib/scan/job-side-effects.test.ts src/lib/scan/scan-session-storage.test.ts src/app/api/extract/route.provider.test.ts src/app/api/extract/route.contract.test.ts src/app/api/scan-jobs/process/route.extractor.test.ts src/app/api/scan-jobs/process/route.contract.test.ts src/app/api/scan-jobs/create/route.contract.test.ts src/app/api/words/create/route.test.ts src/app/api/onboarding/route.test.ts src/app/api/subscription/appstore/verify/route.test.ts src/app/api/subscription/appstore/notifications/route.test.ts src/app/api/shared-projects/shared.test.ts"
+    "test:web": "tsx --test src/lib/api/internal-worker.test.ts src/lib/auth/signup-flow.test.ts src/lib/auth/otp-lifecycle.test.ts src/app/api/auth/otp.contract.test.ts src/lib/stripe/client.test.ts src/lib/appstore/config.test.ts src/lib/appstore/verify.test.ts src/lib/appstore/notifications.test.ts src/lib/subscription/status.test.ts src/lib/subscription/display.test.ts src/lib/subscription/reconcile-status.test.ts src/lib/subscription/billing-activation.test.ts src/lib/subscription/stripe-webhook-handlers.test.ts src/lib/supabase/session-cache.test.ts src/lib/db/remote-repository.test.ts src/lib/db/hybrid-repository.test.ts src/lib/db/sync-queue.test.ts src/lib/db/debounced-status-write.test.ts src/lib/projects/load-helpers.test.ts src/lib/stats/calendar.test.ts src/lib/ai/extract-circled-words.test.ts src/lib/ai/extract-highlighted-words.test.ts src/lib/ai/extract-eiken-words.test.ts src/lib/ai/extract-wrong-answers.test.ts src/lib/ai/generate-example-sentences.test.ts src/lib/ai/match-passage-words.test.ts src/lib/ai/prompts.translation-context.test.ts src/lib/ai/prompts.contract.test.ts src/lib/lexicon/master-first-scan.test.ts src/lib/lexicon/resolver.test.ts src/lib/lexicon/word-resolution-jobs.test.ts src/lib/words/backfill-japanese.test.ts src/lib/schemas/ai-response.test.ts src/lib/scan/mode-provider.test.ts src/lib/scan/job-result-payload.test.ts src/lib/scan/server-cloud-result-payload.test.ts src/lib/scan/quiz-prefill.test.ts src/lib/scan/post-processing.test.ts src/lib/scan/example-generation.test.ts src/lib/scan/image-extraction.test.ts src/lib/scan/server-cloud-persistence.contract.test.ts src/lib/scan/job-side-effects.test.ts src/lib/scan/scan-session-storage.test.ts src/app/api/extract/route.provider.test.ts src/app/api/extract/route.contract.test.ts src/app/api/scan-jobs/process/route.extractor.test.ts src/app/api/scan-jobs/process/route.contract.test.ts src/app/api/scan-jobs/create/route.contract.test.ts src/app/api/words/create/route.test.ts src/app/api/onboarding/route.test.ts src/app/api/subscription/appstore/verify/route.test.ts src/app/api/subscription/appstore/notifications/route.test.ts src/app/api/shared-projects/shared.test.ts"
   },
   "dependencies": {
     "@apple/app-store-server-library": "^2.0.0",
