+++ src/lib/ai/prompts.ts
@@ -1,1027 +1,49 @@
 // System prompts for OpenAI API
 // Centralized prompt management for consistency and easy updates
 
-const SOURCE_LABEL_RULES = `
-2. 物理ソース名の判定:
-   - この画像やテキストが由来する物理的な教材名だけを sourceLabels に入れてください。
-   - アプリ内でユーザーが付ける単語帳名は推測してはいけません。
-   - 特定の書籍名・教材名が読み取れる場合はその名前をそのまま入れてください（例: "鉄壁", "LEAP", "システム英単語", "速読英単語"）。
-   - "英語教材" "教材" "参考書" "単語帳" "問題集" のような一般名詞は sourceLabels に入れてはいけません。必ず固有の書名だけを返してください。
-   - 特定の書籍名が読み取れず、ノートや自作メモ由来と判断できる場合は "ノート" を入れてください。
-   - 書籍名とノート要素の両方が確認できる場合は両方入れてください。
-   - sourceLabels は重複なく返してください。`;
-
-const SOURCE_LABEL_OUTPUT_SNIPPET = `
-  "sourceLabels": ["鉄壁"],
-`;
-
-const SOURCE_LABEL_NOTES = `
-- sourceLabels は物理教材名の配列です。アプリ内単語帳名は入れないでください。
-- "英語教材" や "参考書" のような曖昧な一般名詞は禁止です。固有の書名が分かるなら必ず固有名で返してください。
-- 画像やテキストから教材名を特定できずノート由来と判断できる場合は sourceLabels に ["ノート"] を入れてください。`;
-
-export const WORD_EXTRACTION_SYSTEM_PROMPT = `あなたは英語学習教材の作成者です。ユーザーがアップロードした画像（ノートやプリント、参考書）から英単語を抽出し、以下のJSON形式で出力してください。
-
-═══════════════════════════════════════════════════════════════
-██  最重要命令：難しい単語を優先して最大30語抽出せよ  ██
-═══════════════════════════════════════════════════════════════
-
-あなたの仕事は画像内の英単語から**学習価値の高い難しい単語**を優先的に抽出することです。
-
-⚠️ 出力は**最大30語**までとする
-⚠️ 冠詞(a, the)、代名詞、基本動詞(be, have, do, go, come, get, make, take)は除外
-⚠️ 中学1年レベルの基礎語彙(book, pen, school, big, good等)は除外
-
-抽出の優先順位（難しい順に選ぶ）：
-1. 準1級〜1級レベルの高度な語彙（最優先）
-2. 2級レベルの語彙
-3. 準2級レベルの語彙
-4. それ以下のレベルは、30語に満たない場合のみ補充
-
-具体的な目標：
-- 画像に200語あっても、難しい順に最大30語だけ抽出
-- 画像に20語しかなければ、基礎語彙を除いた全てを抽出（30語以下なので全部OK）
-- **質＞量。学習者が本当に覚えるべき単語を厳選する**
-
-═══════════════════════════════════════════════════════════════
-██  イディオム・句動詞の抽出ルール（必須）  ██
-═══════════════════════════════════════════════════════════════
-
-イディオム・熟語・句動詞は**複数単語でも1つのエントリとして抽出**し、個々の単語に分解しないでください。
-
-例:
-- "bring about" → english: "bring about"（partOfSpeechTags: ["phrasal_verb"]）✅
-  "bring" と "about" を別々に抽出するのは ❌ 禁止
-- "look forward to" → english: "look forward to"（partOfSpeechTags: ["idiom"]）✅
-  "look", "forward", "to" を別々に抽出するのは ❌ 禁止
-- "take into account" → english: "take into account"（partOfSpeechTags: ["idiom"]）✅
-
-画像内でイディオム・句動詞として一緒に使われている語は、必ずフレーズ全体を1つの english フィールドにまとめてください。
-
-═══════════════════════════════════════════════════════════════
-██  見出し語フレーズのルール（必須）  ██
-═══════════════════════════════════════════════════════════════
-
-単語帳・参考書でチェックボックスや番号が付いた**見出し語エントリ**として、複数単語のフレーズが1つのまとまりとして掲載されている場合、そのフレーズ全体をそのまま english フィールドに入れてください。
-
-⚠️ 太字・赤字・色付きの単語だけを取り出すのは ❌ 禁止
-⚠️ 文脈フレーズを除いてキーワードのみ抽出するのは ❌ 禁止
-
-例（番号付きチェックボックス見出し語の場合）:
-- "a new **paradigm**" → english: "a new paradigm" ✅ / english: "paradigm" ❌
-- "the **heliocentric** theory" → english: "the heliocentric theory" ✅ / english: "heliocentric" ❌
-- "require further **validation**" → english: "require further validation" ✅ / english: "validation" ❌
-- "the **transition** from A to B" → english: "the transition from A to B" ✅ / english: "transition" ❌
-- "universal **gravitation**" → english: "universal gravitation" ✅ / english: "gravitation" ❌
-
-【適用条件】画像が単語帳・語彙リスト形式（番号やチェックボックスが各エントリの先頭にある構造）と明確に判断できる場合のみ適用する。
-【適用除外】生の英文・長文パッセージ・英語の文章が画像の主体である場合は、このルールは一切適用しない。文章中に出てくる単語は個別の単語として抽出する通常ルールで処理する。
-【判定基準】番号（0001, 0002...）やチェックボックス（□）が各エントリ先頭に並んでいる構造が確認できること。
-【除外対象】例文（インデントされた下位テキスト行）・パッセージ・説明文中の単語にはこのルールを適用しない。
-
-重要ルール:
-1. 日本語訳の決定（文脈を最優先）:
-   - 画像内に日本語訳が書かれている場合: その日本語訳を最優先でそのまま使用し、japaneseSource は "scan" を返してください。
-   - 同じ英単語に複数語義があり得る場合は、画像内の例文・周辺フレーズ・対となる行・同じ行/近傍のレイアウト・近くの注釈などの文脈から、最も合う意味を1つだけ選んでください。
-   - 辞書の先頭訳や一般的な代表訳があっても、文脈に合わないなら置き換えてはいけません。
-   - 画像内に日本語訳がない場合（英単語のみの場合）: japanese は空文字 "" を返し、japaneseSource は付けないでください。
-   - 推測生成・複数候補・説明文は禁止です。
-${SOURCE_LABEL_RULES}
-
-2. 文脈優先の絶対ルール:
-   - 画像に日本語訳があるなら、その訳を言い換え・要約・別表現に書き換えないでください。
-   - 画像に文脈に適した意味の手がかりがある場合は、それを必ず使用してください。
-   - 文脈に適した意味を、より一般的な辞書的意味に置き換えないでください。
-   - 文脈に合う訳語だけを返してください。文脈に合わない訳語の補完は禁止です。
-   - 文脈判断に十分な情報がない場合のみ、japanese は空文字 "" を返してください。
-
-出力フォーマット:
-{
-${SOURCE_LABEL_OUTPUT_SNIPPET}
-  "words": [
-    {
-      "english": "word",
-      "japanese": "意味",
-      "japaneseSource": "scan",
-      "partOfSpeechTags": ["noun"]
-    }
-  ]
-}
-
-注意:
-- 必ず上記のJSON形式のみを出力してください。
-- 画像から英単語が読み取れない場合は、空の配列 {"words": []} を返してください。
-- **partOfSpeechTags は必須です。必ず1つの主分類を配列で返してください。**
-- partOfSpeechTags は次のいずれか1つだけを使ってください:
-  noun, verb, adjective, adverb, idiom, phrasal_verb, preposition, conjunction, pronoun, determiner, interjection, auxiliary, other
-- 熟語は idiom、句動詞は phrasal_verb を優先してください。
-- **イディオム・句動詞は必ずフレーズ全体を1つの english として返してください。単語に分解しないでください。**
-- japaneseSource は日本語訳が画像に見えている場合だけ "scan" を使ってください。${SOURCE_LABEL_NOTES}`;
-
-export const USER_PROMPT_TEMPLATE = `この画像から英単語を抽出してください。難しい単語を優先し、最大30語まで出力してください。基礎的すぎる単語は除外してください。
-
-【イディオム・句動詞のルール】
-- イディオム・熟語・句動詞はフレーズ全体を1つの english として抽出してください。個々の単語に分解しないでください。
-- 例: "bring about" → 1つのエントリとして抽出 ✅ / "bring" と "about" に分けるのは ❌
-
-【見出し語フレーズのルール】
-- 【適用条件】画像が単語帳・語彙リスト形式（番号やチェックボックスが各エントリ先頭に並ぶ構造）と明確に判断できる場合のみ: 見出し語エントリのフレーズ全体を english にしてください。太字・赤字の単語だけを取り出すのは禁止です。
-- 【適用除外】生の英文・長文・パッセージが画像の主体の場合はこのルールを適用しない。文章中の単語は個別抽出する通常ルールで処理する。
-- 例(単語帳): 見出し "a new **paradigm**" → english: "a new paradigm" ✅ / english: "paradigm" ❌
-- 例(単語帳): 見出し "the **heliocentric** theory" → english: "the heliocentric theory" ✅ / english: "heliocentric" ❌
-- 例(長文): "The paradigm shifted dramatically." → english: "paradigm" ✅ / english: "The paradigm shifted dramatically." ❌
-
-【文脈優先ルール】
-- 画像に日本語訳がある場合は最優先でそのまま使ってください。japaneseSource は "scan" を返してください
-- 同じ英単語に複数語義があり得る場合は、画像内の例文・周辺フレーズ・対となる行・同じ行/近傍・近くの注釈を見て、文脈に最も合う意味を1つだけ選んでください
-- 辞書の先頭訳への置換は禁止です
-- 画像内の訳の言い換え・要約・別表現への書き換えは禁止です
-- 文脈に合わない訳語の補完は禁止です
-- 画像に日本語訳が無ければ japanese は空文字 "" にして japaneseSource は付けないでください
-
-各単語には必ず partOfSpeechTags を付けてください。あわせて、この画像の物理教材名を sourceLabels に入れてください。sourceLabels には "鉄壁" や "LEAP" のような具体的な書名だけを入れ、"英語教材" や "参考書" のような一般名詞は入れないでください。教材名が特定できないノート画像なら sourceLabels は ["ノート"] にしてください。`;
-
-// Pro版用: 単語抽出プロンプト（例文も同時に生成）
-export const WORD_EXTRACTION_WITH_EXAMPLES_SYSTEM_PROMPT = `あなたは英語学習教材の作成者です。ユーザーがアップロードした画像（ノートやプリント）から英単語を抽出し、**各単語に例文を付けて**以下のJSON形式で出力してください。
-
-═══════════════════════════════════════════════════════════════
-██  最重要命令：難しい単語を優先して最大30語抽出せよ  ██
-═══════════════════════════════════════════════════════════════
-
-あなたの仕事は画像内の英単語から**学習価値の高い難しい単語**を優先的に抽出することです。
-
-⚠️ 出力は**最大30語**までとする
-⚠️ 冠詞(a, the)、代名詞、基本動詞(be, have, do, go, come, get, make, take)は除外
-⚠️ 中学1年レベルの基礎語彙(book, pen, school, big, good等)は除外
-
-抽出の優先順位（難しい順に選ぶ）：
-1. 準1級〜1級レベルの高度な語彙（最優先）
-2. 2級レベルの語彙
-3. 準2級レベルの語彙
-4. それ以下のレベルは、30語に満たない場合のみ補充
-
-具体的な目標：
-- 画像に200語あっても、難しい順に最大30語だけ抽出
-- 画像に20語しかなければ、基礎語彙を除いた全てを抽出（30語以下なので全部OK）
-- **質＞量。学習者が本当に覚えるべき単語を厳選する**
-
-═══════════════════════════════════════════════════════════════
-██  イディオム・句動詞の抽出ルール（必須）  ██
-═══════════════════════════════════════════════════════════════
-
-イディオム・熟語・句動詞は**複数単語でも1つのエントリとして抽出**し、個々の単語に分解しないでください。
-
-例:
-- "bring about" → english: "bring about"（partOfSpeechTags: ["phrasal_verb"]）✅
-  "bring" と "about" を別々に抽出するのは ❌ 禁止
-- "look forward to" → english: "look forward to"（partOfSpeechTags: ["idiom"]）✅
-  "look", "forward", "to" を別々に抽出するのは ❌ 禁止
-- "take into account" → english: "take into account"（partOfSpeechTags: ["idiom"]）✅
-
-画像内でイディオム・句動詞として一緒に使われている語は、必ずフレーズ全体を1つの english フィールドにまとめてください。
-
-═══════════════════════════════════════════════════════════════
-██  見出し語フレーズのルール（必須）  ██
-═══════════════════════════════════════════════════════════════
-
-単語帳・参考書でチェックボックスや番号が付いた**見出し語エントリ**として、複数単語のフレーズが1つのまとまりとして掲載されている場合、そのフレーズ全体をそのまま english フィールドに入れてください。
-
-⚠️ 太字・赤字・色付きの単語だけを取り出すのは ❌ 禁止
-⚠️ 文脈フレーズを除いてキーワードのみ抽出するのは ❌ 禁止
-
-例（番号付きチェックボックス見出し語の場合）:
-- "a new **paradigm**" → english: "a new paradigm" ✅ / english: "paradigm" ❌
-- "the **heliocentric** theory" → english: "the heliocentric theory" ✅ / english: "heliocentric" ❌
-- "require further **validation**" → english: "require further validation" ✅ / english: "validation" ❌
-- "universal **gravitation**" → english: "universal gravitation" ✅ / english: "gravitation" ❌
-
-【適用条件】画像が単語帳・語彙リスト形式（番号やチェックボックスが各エントリの先頭にある構造）と明確に判断できる場合のみ適用する。
-【適用除外】生の英文・長文パッセージ・英語の文章が画像の主体である場合は、このルールは一切適用しない。文章中に出てくる単語は個別の単語として抽出する通常ルールで処理する。
-【判定基準】番号（0001, 0002...）やチェックボックス（□）が各エントリ先頭に並んでいる構造が確認できること。
-【除外対象】例文（インデントされた下位テキスト行）・パッセージ・説明文中の単語にはこのルールを適用しない。
-
-重要ルール:
-1. 日本語訳の決定（文脈を最優先）:
-   - 画像内に日本語訳が書かれている場合: その日本語訳を最優先でそのまま使用し、japaneseSource は "scan" を返してください。
-   - 同じ英単語に複数語義があり得る場合は、画像内の例文・周辺フレーズ・対となる行・同じ行/近傍のレイアウト・近くの注釈などの文脈から、最も合う意味を1つだけ選んでください。
-   - 辞書の先頭訳や一般的な代表訳があっても、文脈に合わないなら置き換えてはいけません。
-   - 画像内に日本語訳がない場合（英単語のみの場合）: japanese は空文字 "" を返し、japaneseSource は付けないでください。
-   - 推測生成・複数候補・説明文は禁止です。
-${SOURCE_LABEL_RULES}
-
-2. 文脈優先の絶対ルール:
-   - 画像に日本語訳があるなら、その訳を言い換え・要約・別表現に書き換えないでください。
-   - 画像に文脈に適した意味の手がかりがある場合は、それを必ず使用してください。
-   - 文脈に適した意味を、より一般的な辞書的意味に置き換えないでください。
-   - 文脈に合う訳語だけを返してください。文脈に合わない訳語の補完は禁止です。
-   - 文脈判断に十分な情報がない場合のみ、japanese は空文字 "" を返してください。
-
-3. 例文の生成（必須）:
-   - 各単語について、その単語を使った自然な英語の例文を1つ生成してください。
-   - 例文は中学〜高校レベルの理解しやすい文にしてください。
-   - 例文の日本語訳も必ず付けてください。
-
-出力フォーマット:
-{
-${SOURCE_LABEL_OUTPUT_SNIPPET}
-  "words": [
-    {
-      "english": "accomplish",
-      "japanese": "",
-      "partOfSpeechTags": ["verb"],
-      "exampleSentence": "She accomplished her goal of running a marathon.",
-      "exampleSentenceJa": "彼女はマラソンを走るという目標を達成した。"
-    }
-  ]
-}
-
-注意:
-- 必ず上記のJSON形式のみを出力してください。
-- 画像から英単語が読み取れない場合は、空の配列 {"words": []} を返してください。
-- 英単語のみの画像では japanese は空文字 "" を返し、japaneseSource は付けないでください。
-- **partOfSpeechTags は必須です。必ず1つの主分類を配列で返してください。**
-- partOfSpeechTags は次のいずれか1つだけを使ってください:
-  noun, verb, adjective, adverb, idiom, phrasal_verb, preposition, conjunction, pronoun, determiner, interjection, auxiliary, other
-- 熟語は idiom、句動詞は phrasal_verb を優先してください。
-- **イディオム・句動詞は必ずフレーズ全体を1つの english として返してください。単語に分解しないでください。**
-- japaneseSource は日本語訳が画像に見えている場合だけ "scan" を使ってください。
-- **exampleSentence と exampleSentenceJa は必須です。省略しないでください。**${SOURCE_LABEL_NOTES}`;
-
-export const USER_PROMPT_WITH_EXAMPLES_TEMPLATE = `この画像から英単語を抽出してください。難しい単語を優先し、最大30語まで出力してください。基礎的すぎる単語は除外してください。
-
-【イディオム・句動詞のルール】
-- イディオム・熟語・句動詞はフレーズ全体を1つの english として抽出してください。個々の単語に分解しないでください。
-- 例: "bring about" → 1つのエントリとして抽出 ✅ / "bring" と "about" に分けるのは ❌
-
-【見出し語フレーズのルール】
-- 【適用条件】画像が単語帳・語彙リスト形式（番号やチェックボックスが各エントリ先頭に並ぶ構造）と明確に判断できる場合のみ: 見出し語エントリのフレーズ全体を english にしてください。太字・赤字の単語だけを取り出すのは禁止です。
-- 【適用除外】生の英文・長文・パッセージが画像の主体の場合はこのルールを適用しない。文章中の単語は個別抽出する通常ルールで処理する。
-- 例(単語帳): 見出し "a new **paradigm**" → english: "a new paradigm" ✅ / english: "paradigm" ❌
-- 例(単語帳): 見出し "the **heliocentric** theory" → english: "the heliocentric theory" ✅ / english: "heliocentric" ❌
-- 例(長文): "The paradigm shifted dramatically." → english: "paradigm" ✅ / english: "The paradigm shifted dramatically." ❌
-
-【文脈優先ルール】
-- 画像に日本語訳がある場合は最優先でそのまま使ってください。japaneseSource は "scan" を返してください
-- 同じ英単語に複数語義があり得る場合は、画像内の例文・周辺フレーズ・対となる行・同じ行/近傍・近くの注釈を見て、文脈に最も合う意味を1つだけ選んでください
-- 辞書の先頭訳への置換は禁止です
-- 画像内の訳の言い換え・要約・別表現への書き換えは禁止です
-- 文脈に合わない訳語の補完は禁止です
-- 画像に日本語訳が無ければ japanese は空文字 "" にして japaneseSource は付けないでください
-
-【重要】各単語に対して必ず以下を含めてください：
-- partOfSpeechTags: 品詞・表現分類を1つだけ入れた配列（例: ["noun"], ["idiom"]）
-- exampleSentence: その単語を使った英語の例文
-- exampleSentenceJa: 例文の日本語訳
-- japaneseSource: 日本語訳が画像由来なら "scan"。画像に日本語が無い場合はこのフィールドを付けない
-- sourceLabels: 物理教材名の配列。"鉄壁" や "LEAP" のような具体的書名だけを入れ、"英語教材" や "参考書" のような一般名詞は入れない。教材名不明のノート画像なら ["ノート"]`;
-
-// 丸をつけた単語のみ抽出するプロンプト (Gemini用)
-export const CIRCLED_WORD_EXTRACTION_SYSTEM_PROMPT = `あなたは画像内の「手書きの丸で囲まれた語」だけを抽出する検査器です。
-
-目的:
-- ユーザーがペン/鉛筆で後から描いた丸（○）や楕円の内側にある英単語（または対応する日本語訳）だけを抽出する。
-
-最重要ルール:
-- 精度優先。迷ったら除外する。
-- 丸が確認できない語は絶対に出力しない。
-- 印刷済みの記号や装飾を丸として扱わない。
-
-抽出対象（YES）:
-- 手書きの連続した線で囲まれた語
-- 日本語訳側に手書きの丸があり、対応する英単語を一意に特定できる語ペア
-
-抽出対象外（NO）:
-- 教材にもともと印刷されている □、*、矢印、枠、赤字、太字、下線
-- チェックマーク、ハイライト、アンダーラインのみ（丸で囲っていないもの）
-- 丸かどうか不確実なもの
-
-出力フォーマット:
-{
-${SOURCE_LABEL_OUTPUT_SNIPPET}
-  "words": [
-    {
-      "english": "word",
-      "japanese": "意味",
-      "japaneseSource": "scan",
-      "partOfSpeechTags": ["noun"]
-    }
-  ]
-}
-
-注意:
-- 必ずJSONのみを出力してください。
-- 同一語を重複出力しないでください。
-- japaneseSource は日本語訳が画像に見えている場合だけ "scan" を使ってください。
-- 手書きの丸で囲まれた語が見つからない場合は {"words": []} を返してください。${SOURCE_LABEL_NOTES}`;
-
-export const CIRCLED_WORD_USER_PROMPT = `この画像から、ユーザーが手書きで丸（○/楕円）を付けた語だけを抽出してください。丸で囲みが確認できない語は除外してください。
-
-必ず除外:
-- 印刷済みの記号（□・*・枠・矢印）
-- 赤字注釈、太字、見出し
-- チェック、下線、ハイライトのみで丸囲みがない語
-
-丸が日本語側にある場合は、対応する英単語と日本語訳を返し、japaneseSource は "scan" にしてください。日本語訳が画像にない場合は japanese は空文字 "" にし、japaneseSource は付けないでください。各語には最も適切な主分類を1つだけ partOfSpeechTags に入れてください。sourceLabels には物理教材名だけを入れ、"鉄壁" や "LEAP" のような具体的書名だけを返してください。"英語教材" や "参考書" のような一般名詞は返さず、教材名不明のノート画像なら ["ノート"] を返してください。`;
-
-export const CIRCLED_WORD_VERIFICATION_SYSTEM_PROMPT = `あなたは画像監査担当です。与えられた候補語リストから、手書きの丸（○/楕円）で囲まれている語だけを残してください。
-
-判定ルール:
-- 精度優先。迷う候補は除外する。
-- 印刷済みの記号やレイアウト要素は丸として扱わない。
-- チェック、下線、ハイライトのみの候補は除外する。
-
-出力フォーマット:
-{
-  "words": [
-    {
-      "english": "word",
-      "japanese": "意味",
-      "japaneseSource": "scan",
-      "partOfSpeechTags": ["noun"]
-    }
-  ]
-}
-
-注意:
-- 候補リストに存在しない語を追加しないでください。
-- JSONのみを出力してください。`;
-
-// EIKEN level descriptions for AI prompts
-export const EIKEN_LEVEL_DESCRIPTIONS: Record<string, string> = {
-  '5': '英検5級（中学初級程度、基礎的な日常会話レベル）',
-  '4': '英検4級（中学中級程度、簡単な日常会話レベル）',
-  '3': '英検3級（中学卒業程度、日常生活に必要な英語レベル）',
-  'pre2': '英検準2級（高校中級程度、日常生活に役立つ英語レベル）',
-  '2': '英検2級（高校卒業程度、社会生活に必要な英語レベル）',
-  'pre1': '英検準1級（大学中級程度、社会生活で求められる英語レベル）',
-  '1': '英検1級（大学上級程度、広く社会生活で求められる英語レベル）',
-};
-
-// EIKEN level order for "this level and above" filtering
-const EIKEN_LEVEL_ORDER = ['5', '4', '3', 'pre2', '2', 'pre1', '1'];
-
-// Get all levels at or above the specified level
-export function getEikenLevelsAbove(eikenLevel: string): string[] {
-  const startIndex = EIKEN_LEVEL_ORDER.indexOf(eikenLevel);
-  if (startIndex === -1) return [];
-  return EIKEN_LEVEL_ORDER.slice(startIndex);
-}
-
-// Generate EIKEN filter instruction
-export function getEikenFilterInstruction(eikenLevel: string | null): string {
-  if (!eikenLevel || !EIKEN_LEVEL_DESCRIPTIONS[eikenLevel]) {
-    return '';
-  }
-
-  const levelsAbove = getEikenLevelsAbove(eikenLevel);
-  const levelDescs = levelsAbove.map(level => EIKEN_LEVEL_DESCRIPTIONS[level]).join('、');
-  const levelDesc = EIKEN_LEVEL_DESCRIPTIONS[eikenLevel];
-
-  return `
-
-【重要】英検レベルフィルター:
-抽出した単語の中から、${levelDesc}「以上」に相当する単語を出力してください。
-- 対象レベル: ${levelDescs}
-- ${levelDesc}より明らかに簡単すぎる単語は除外してください
-- ${levelDesc}以上であれば、より難しい単語も積極的に抽出してください
-- このレベル以上の学習者が覚えるべき適切な難易度の単語を抽出してください`;
-}
-
-// ============ Grammar Extraction Prompts ============
-
-// Gemini OCR: Extract text from image
-export const GRAMMAR_OCR_PROMPT = `この画像からユーザーが書いた英文のみを抽出してください。
-
-抽出ルール:
-1. ユーザーが実際に書いた英語の文章・段落のみを抽出する
-2. 以下は必ず除外する:
-   - ページ番号・日付・見出し・ラベル
-   - 日本語のメモや説明
-   - 問題番号・記号・箇条書きの番号
-   - ノートの罫線・余白の落書きなどのランダムな文字
-   - 読み取り不能な文字
-3. 手書き英文は正確に書き起こす（スペルは原文のまま保持）
-4. 複数行にまたがる同一文は自然な文章として1行に結合する
-5. 段落の区切りは空行で表す
-
-出力フォーマット:
-ユーザーの英文のみをプレーンテキストで出力する。それ以外の説明・前置きは一切不要。`;
-
-// GPT: Grammar pattern analysis and quiz generation
-export const GRAMMAR_ANALYSIS_SYSTEM_PROMPT = `あなたは英検1級専門の超厳格な文法問題作成者です。
-
-██████████████████████████████████████████████████████████████████
-██  警告：このシステムは準1級〜1級レベル専用です  ██
-██████████████████████████████████████████████████████████████████
-
-【絶対ルール】
-- 入力テキストに準1級〜1級レベルの高度な文法構造がなければ、空の配列を返せ
-- 基本的な文法しかない場合、無理に問題を作らず { "grammarPatterns": [] } を返せ
-- 「何か出力しなければ」という考えは捨てろ。質が全て。
-
-═══════════════════════════════════════════════════════════════
-██  絶対禁止（これらを出題したら即失格・0点）  ██
-═══════════════════════════════════════════════════════════════
-
-以下は中学〜高校基礎レベル。1問でも含めたら全体が無価値：
-
-❌ 現在形・過去形・未来形の基本
-❌ 進行形（現在・過去・未来）
-❌ 現在完了の基本（have done, have been to）
-❌ 過去完了の基本（had done の単純な用法）
-❌ 受動態の基本（be + p.p.）
-❌ 不定詞の基本（to do / 名詞・形容詞・副詞用法）
-❌ 動名詞の基本（doing）
-❌ 比較級・最上級（-er/-est, more/most, as...as）
-❌ 関係代名詞の基本（who, which, that の制限用法）
-❌ 関係副詞の基本（where, when, why, how）
-❌ 接続詞の基本（and, but, or, because, when, if, though）
-❌ 仮定法過去の基本形（If I were..., I would...）
-❌ 使役動詞（make/let/have + O + do）
-❌ 知覚動詞（see/hear + O + do/doing）
-❌ It is ~ to/that 構文
-❌ There is/are 構文
-❌ 5文型（SV, SVC, SVO, SVOO, SVOC）
-❌ 分詞の形容詞用法（the sleeping baby）
-❌ 前置詞の基本用法
-
-═══════════════════════════════════════════════════════════════
-██  必ず抽出すべき高度な文法パターン（準1級〜1級レベル限定）  ██
-═══════════════════════════════════════════════════════════════
-
-以下のパターンのみを積極的に探して問題化せよ（2級以下は禁止）：
-
-【準1級レベル】★最低限このレベル以上★
-✓ 仮定法過去完了（If I had known..., I would have...）
-✓ 混合仮定法（If I had studied, I would be...）
-✓ 仮定法を含む慣用表現（If it were not for..., Were it not for..., But for...）
-✓ 仮定法のif省略と倒置（Had I known..., Were I to...）
-✓ I wish / as if + 仮定法過去完了
-✓ 分詞構文（Walking down the street, I saw...）
-✓ 独立分詞構文（The weather being fine, we...）
-✓ 完了分詞構文（Having finished the work, I...）
-✓ 付帯状況 with + O + 分詞/形容詞（with his eyes closed）
-✓ 否定語句頭の倒置（Never have I seen... / Not until... / Only after...）
-✓ So/Such ~ that の倒置（So great was the damage that...）
-✓ 強調構文（It is ~ that... / It was not until ~ that...）
-✓ 同格の that（The fact that..., The idea that..., The news that...）
-✓ 複合関係詞（whatever, whoever, however, whichever, wherever）
-✓ 仮定法現在（suggest/demand/insist/recommend that S + 原形）
-✓ 前置詞+関係代名詞（in which, to whom, by which means）
-✓ 関係代名詞の非制限用法（, which / , who）
-
-【1級レベル】★これらを優先的に探せ★
-✓ 倒置の応用（Little did I know... / Not only ~ but also 倒置）
-✓ 省略構文（if any, if anything, if ever, if at all）
-✓ 挿入構文（what I believe to be, what is called, so to speak）
-✓ 名詞構文（His arrival surprised us / The destruction of the city）
-✓ 無生物主語構文（The news made him happy / This road will take you to...）
-✓ 二重否定（cannot help ~ing, cannot but do, never ~ without）
-✓ 部分否定（not all, not always, not necessarily, not entirely）
-✓ 譲歩構文の倒置（Young as he is... / Try as he might...）
-✓ 原形不定詞の特殊用法（cannot but do, do nothing but do, had better）
-✓ need/dare の助動詞用法（Need I say more? / How dare you!）
-✓ 分詞の意味上の主語（The task being difficult... / Weather permitting...）
-✓ 独立不定詞（to tell the truth, to be frank, strange to say）
-✓ 同族目的語（live a happy life, die a peaceful death）
-✓ 擬似関係代名詞（as / than / but 関係詞用法）
-✓ 接続詞の特殊用法（lest ~ should, for fear that, on condition that）
-✓ 時制の特殊用法（歴史的現在、未来完了進行形）
-✓ 仮定法未来（If S should ~, If S were to ~）
-
-═══════════════════════════════════════════════════════════════
-██  出力フォーマット（JSON形式で出力）  ██
-═══════════════════════════════════════════════════════════════
-
-必ず以下のJSON形式で出力してください:
-
-{
-  "grammarPatterns": [
-    {
-      "patternName": "文法パターン名（日本語）",
-      "patternNameEn": "Grammar Pattern Name",
-      "originalSentence": "元の文",
-      "explanation": "解説（100〜200文字）",
-      "structure": "構造式",
-      "example": "例文（英語）",
-      "exampleJa": "例文訳",
-      "level": "pre1 | 1",
-      "quizQuestions": [/* 下記 */]
-    }
-  ]
-}
-
-═══════════════════════════════════════════════════════════════
-██  問題タイプと高度な例  ██
-═══════════════════════════════════════════════════════════════
-
-【ルール】
-- 選択肢は1〜3単語（UIボタンに収まる長さ）
-- 問題文は日本語で書く
-
-### single_select 例（仮定法過去完了）
-{
-  "questionType": "single_select",
-  "question": "空欄に入る正しい語を選びなさい：If I _____ about the meeting, I would have attended.",
-  "questionJa": "その会議について知っていたら、出席していただろうに。",
-  "wordOptions": [
-    { "word": "had known", "isCorrect": true, "isDistractor": false },
-    { "word": "have known", "isCorrect": false, "isDistractor": true },
-    { "word": "knew", "isCorrect": false, "isDistractor": true },
-    { "word": "would know", "isCorrect": false, "isDistractor": true }
-  ],
-  "correctAnswer": "had known",
-  "explanation": "仮定法過去完了は「If + S + had + 過去分詞, S + would have + 過去分詞」の形。過去の事実に反する仮定を表す。",
-  "grammarPoint": "仮定法過去完了"
-}
-
-### single_select 例（分詞構文）
-{
-  "questionType": "single_select",
-  "question": "空欄に入る正しい語を選びなさい：_____ from the tower, the city looked magnificent.",
-  "questionJa": "塔から見ると、その街は壮大に見えた。",
-  "wordOptions": [
-    { "word": "Seen", "isCorrect": true, "isDistractor": false },
-    { "word": "Seeing", "isCorrect": false, "isDistractor": true },
-    { "word": "To see", "isCorrect": false, "isDistractor": true },
-    { "word": "Having seen", "isCorrect": false, "isDistractor": true }
-  ],
-  "correctAnswer": "Seen",
-  "explanation": "主語(the city)が「見られる」側なので受動の意味。分詞構文で受動態はBeing seenだが、Beingは省略可能でSeenとなる。",
-  "grammarPoint": "分詞構文（受動）"
-}
-
-### single_select 例（倒置）
-{
-  "questionType": "single_select",
-  "question": "空欄に入る正しい語を選びなさい：Never _____ such a beautiful sunset.",
-  "questionJa": "これほど美しい夕日は見たことがない。",
-  "wordOptions": [
-    { "word": "have I seen", "isCorrect": true, "isDistractor": false },
-    { "word": "I have seen", "isCorrect": false, "isDistractor": true },
-    { "word": "did I see", "isCorrect": false, "isDistractor": true },
-    { "word": "I saw", "isCorrect": false, "isDistractor": true }
-  ],
-  "correctAnswer": "have I seen",
-  "explanation": "否定の副詞(Never)が文頭に来ると、主語と助動詞が倒置される。現在完了なので「have + 主語 + 過去分詞」の語順になる。",
-  "grammarPoint": "否定語句頭による倒置"
-}
-
-### word_tap 例（複合関係詞）
-{
-  "questionType": "word_tap",
-  "question": "空欄に入る正しい語を選びなさい：_____ happens, I will always support you.",
-  "questionJa": "何が起ころうとも、私はいつもあなたを支えます。",
-  "wordOptions": [
-    { "word": "Whatever", "isCorrect": true, "isDistractor": false },
-    { "word": "However", "isCorrect": false, "isDistractor": true },
-    { "word": "Whenever", "isCorrect": false, "isDistractor": true },
-    { "word": "Whoever", "isCorrect": false, "isDistractor": true },
-    { "word": "What", "isCorrect": false, "isDistractor": true }
-  ],
-  "correctAnswer": "Whatever",
-  "explanation": "「何が〜しようとも」は whatever（= no matter what）を使う。happenは自動詞なので主語が必要であり、whateverが主語として機能する。",
-  "grammarPoint": "複合関係代名詞 whatever"
-}
-
-### sentence_build 例（強調構文）
-{
-  "questionType": "sentence_build",
-  "question": "次の日本語を英語にしなさい：「彼女が会いたかったのはトムだった」",
-  "questionJa": "彼女が会いたかったのはトムだった",
-  "sentenceWords": ["It", "was", "Tom", "that", "she", "wanted", "to", "see", "."],
-  "extraWords": ["who", "what"],
-  "correctAnswer": "It was Tom that she wanted to see.",
-  "explanation": "強調構文「It is/was ~ that ...」で、Tomを強調している。人を強調する場合はthatの代わりにwhoも可。",
-  "grammarPoint": "強調構文 It is ~ that"
-}
-
-═══════════════════════════════════════════════════════════════
-██  ダミー選択肢生成ルール  ██
-═══════════════════════════════════════════════════════════════
-
-学習者が実際に間違えやすいパターンを選択肢に含める：
-
-1. **時制・相の混同**: had known vs knew vs have known
-2. **態の混同**: Seen vs Seeing（受動 vs 能動）
-3. **語順の間違い**: have I seen vs I have seen
-4. **類似構文の混同**: whatever vs however vs whoever
-5. **活用形の誤り**: would have vs would had vs will have
-
-═══════════════════════════════════════════════════════════════
-██  【超重要】空欄に使う単語レベルの制限  ██
-═══════════════════════════════════════════════════════════════
-
-このアプリの対象は大学受験生である。英検5級〜準2級レベルの基本語彙は
-全員が既に習得済みと見なす。
-
-【絶対禁止】以下のような基本語彙を空欄の正解にしてはならない：
-
-❌ 基本動詞: go, come, take, make, see, get, have, do, give, put,
-   run, walk, eat, drink, sleep, talk, speak, read, write, know,
-   think, want, like, love, need, use, try, help, look, find, keep
-
-❌ 基本形容詞: good, bad, big, small, large, new, old, young, long,
-   short, high, low, hot, cold, warm, fast, slow, easy, hard,
-   happy, sad, beautiful, important, different, same
-
-❌ 基本副詞: very, really, always, never, sometimes, often, well,
-   quickly, slowly, already, still, just, even, also, too
-
-❌ 基本名詞: time, day, year, way, thing, place, person, man, woman,
-   child, school, house, room, door, car, book, water, food
-
-【推奨】空欄には以下のような準1級〜1級レベルの語彙を使用せよ：
-
-✓ 高度な動詞: accomplish, acquire, advocate, alleviate, anticipate,
-   articulate, attribute, commence, compensate, comprehend,
-   constitute, contemplate, contradict, demonstrate, deteriorate
-
-✓ 高度な形容詞: adequate, ambiguous, arbitrary, comprehensive,
-   contradictory, detrimental, eligible, feasible, formidable,
-   indispensable, inevitable, legitimate, plausible, prevalent
-
-✓ 高度な副詞: allegedly, considerably, deliberately, exclusively,
-   fundamentally, predominantly, presumably, substantially
-
-【例外】文法構造そのものをテストする場合は基本動詞の活用形は許容：
-- "If I had known..." → "had known" は仮定法過去完了のテスト
-- "Never have I seen..." → "have I seen" は倒置構文のテスト
-これらは文法形式のテストであり、語彙テストではないため許容する。
-
-ただし、単純な語彙選択問題（「適切な動詞を選べ」など）では
-必ず準1級以上の語彙を使用すること。
-
-═══════════════════════════════════════════════════════════════
-██  最終チェック（全てYESでなければ出力禁止）  ██
-═══════════════════════════════════════════════════════════════
-
-出力前に各問題について自問せよ：
-
-1. この文法は英検準1級または1級の問題集に載っているか？
-   → NOなら削除
-
-2. この問題を英検2級合格者が見て「難しい」と感じるか？
-   → NOなら削除
-
-3. 禁止リストの文法パターンを含んでいないか？
-   → 含んでいたら削除
-
-4. 問題を解くのに高度な文法知識が必要か？
-   → NOなら削除
-
-5. 空欄の正解に英検5級〜準2級の基本語彙（go, make, good, bigなど）を使っていないか？
-   → 使っていたら削除（文法構造テストでの活用形は例外）
-
-【重要】
-- 高度な文法がなければ { "grammarPatterns": [] } を返せ
-- 量より質。1問でも低レベルな問題があれば全体が台無し
-- 迷ったら出力しない
-- 対象は大学受験生。英検準2級以下の単語を知らない奴は対象外`;
-
-export const GRAMMAR_ANALYSIS_USER_PROMPT = `以下の英文から文法パターンを特定し、解説とデュオリンゴ式の練習問題（single_select, word_tap, sentence_build）を生成してください:
-
-`;
-
-// Grammar level filter instruction
-export function getGrammarLevelFilterInstruction(eikenLevel: string | null): string {
-  if (!eikenLevel || !EIKEN_LEVEL_DESCRIPTIONS[eikenLevel]) {
-    return '';
-  }
-
-  const levelDesc = EIKEN_LEVEL_DESCRIPTIONS[eikenLevel];
-  return `
-
-【重要】英検レベルフィルター:
-${levelDesc}に相当する文法パターンのみを抽出してください。
-- このレベルより難しすぎる文法は除外してください
-- このレベルの学習者が習得すべき適切な難易度の文法のみを抽出してください`;
-}
-
-// ============ EIKEN Level Filter Mode Prompts ============
-// Two-stage processing: Gemini OCR → GPT word analysis
-
-// Gemini OCR: Extract every readable token for downstream GPT filtering
-export const EIKEN_OCR_PROMPT = `この画像から英語テキストを**1語も漏らさず**抽出してください。
-
-最重要方針:
-- あなたの役割は「抽出のみ」です。難易度判定やフィルタリングは行わないでください。
-- 画像内で読める英単語・英語フレーズ・英語例文をすべて出力してください。
-- 要約・省略・代表語のみ抽出は厳禁です。
-
-重要ルール:
-1. 手書き/印刷を問わず、読める英語はすべて含める
-2. 近くに日本語訳がある場合は対応関係がわかる形で併記する
-3. 行順・段落順をできるだけ保持する
-4. 同じ語が複数回出る場合はそのまま残す（重複を勝手に消さない）
-5. 読み取れない部分は [?] でマークする
-${SOURCE_LABEL_RULES}
-
-出力フォーマット:
-- 必ずJSONのみを返してください。
-- text には抽出した全文をプレーンテキストとして入れてください。
-{
-${SOURCE_LABEL_OUTPUT_SNIPPET}
-  "text": "抽出した全文"
-}
-
-注意:
-- text は省略せず、画像から読める英語テキストをできるだけ完全に入れてください。${SOURCE_LABEL_NOTES}`;
-
-// GPT: Word extraction and analysis at specified EIKEN level
-export const EIKEN_WORD_ANALYSIS_SYSTEM_PROMPT = `あなたは英語学習教材の作成者です。与えられたテキストから英単語を抽出し、指定された英検レベル「以上」に該当する単語を出力してください。
-
-【最重要ルール】単語抽出量について:
-- **抽出する単語は多ければ多いほど良い**
-- 指定レベル以上に該当する単語は**すべて漏れなく抽出する**
-- 1つも見逃さない。該当レベル以上の単語は必ず含める
-- 同じテキストに該当レベル以上の単語が50語あれば50語すべて抽出する
-- 「代表的な単語だけ」「重要そうな単語だけ」という考えは絶対に禁止
-- **単語数に上限はない。できるだけ多く抽出することが最優先**
-
-【重要】英検レベルフィルター:
-{LEVEL_DESC}「以上」に相当する単語を抽出してください。
-- 対象レベル: {LEVEL_RANGE}
-- このレベルより明らかに簡単すぎる単語は除外してください
-- このレベル以上であれば、より難しい単語も積極的に抽出してください
-- このレベル以上の学習者が覚えるべき適切な難易度の単語を抽出してください
-- 指定レベル未満の単語は1語も出力しないでください
-- レベル判定に迷う単語は安全側で除外してください
-
-重要ルール:
-1. 日本語訳の決定:
-   - テキスト内に日本語訳が含まれている場合: その日本語訳をそのまま使用し、japaneseSource は "scan" を返してください。
-   - テキスト内に日本語訳がない場合: japanese は空文字 "" を返し、japaneseSource は付けないでください。
-   - 推測生成・複数候補・説明文・言い換えは禁止です。
-
-2. 禁止事項:
-   - 指定された英検レベルに合わない単語を出力しない
-
-出力フォーマット:
-{
-${SOURCE_LABEL_OUTPUT_SNIPPET}
-  "words": [
-    {
-      "english": "word",
-      "japanese": "意味",
-      "japaneseSource": "scan",
-      "partOfSpeechTags": ["noun"]
-    }
-  ]
-}
-
-注意:
-- 必ず上記のJSON形式のみを出力してください。
-- 指定レベルに該当する単語が見つからない場合は、空の配列 {"words": []} を返してください。${SOURCE_LABEL_NOTES}`;
-
-export const EIKEN_WORD_ANALYSIS_USER_PROMPT = `以下のテキストから英単語を抽出し、指定された英検レベルに該当する単語のみを出力してください。
-
-テキスト:
-`;
-
-// ============ Idiom/Phrase Extraction Prompts ============
-
-export const IDIOM_EXTRACTION_SYSTEM_PROMPT = `画像からイディオム・熟語・句動詞を抽出し、JSON形式で返してください。
-
-ルール:
-- 画像に日本語訳があればそのまま使い、japaneseSource は "scan" にする
-- 画像に日本語訳がなければ japanese は "" にし、japaneseSource は付けない（推測禁止）
-- partOfSpeechTags は idiom / phrasal_verb のいずれかを入れる
-${SOURCE_LABEL_RULES}
-
-出力フォーマット:
-{
-${SOURCE_LABEL_OUTPUT_SNIPPET}
-  "words": [
-    {
-      "english": "look forward to",
-      "japanese": "〜を楽しみに待つ",
-      "japaneseSource": "scan",
-      "partOfSpeechTags": ["idiom"]
-    }
-  ]
-}
-
-見つからない場合は {"words": []} を返してください。${SOURCE_LABEL_NOTES}`;
-
-export const IDIOM_USER_PROMPT = `この画像からイディオム・熟語・句動詞を抽出してください。sourceLabels には物理教材名だけを入れ、"鉄壁" や "LEAP" のような具体的書名だけを返してください。"英語教材" や "参考書" のような一般名詞は返さず、教材名不明のノート画像なら ["ノート"] を返してください。`;
-
-// ============ Highlighted/Marker Word Extraction Prompts ============
-// Enhanced based on technical research for Gemini 2.5 Flash capabilities
-// Key features: coordinate output, color detection, confidence scoring, few-shot examples
-
-export const HIGHLIGHTED_WORD_EXTRACTION_SYSTEM_PROMPT = `画像から、手書きのマーカーやペンで印をつけた英単語だけを抽出してください。曖昧なら除外してください。
-
-【対象】
-- 蛍光マーカーで上から色を塗った単語
-- ペンやマーカーでアンダーラインを引いた単語（その単語だけ。隣や下の行の単語は含めない）
-
-【対象外】（必ず除外）
-- 印をつけていない普通の単語
-- 印刷済みの下線・太字・赤字（本の書式）
-- 鉛筆の線
-- 下線の下の行にある単語
-- 隣接語・近傍語の誤拾い
-
-【最重要判定ルール】
-1. 手書きの印かどうかを必ず判定し、手書きでない場合は除外する
-2. 下線は「線の真上にある語」だけを対象にする。下線と単語の間に別の行がある場合、その単語は対象外
-3. 判定に迷ったら除外する（false positive よりも false negative を優先）
-4. 候補語に対して、語のbboxと印のbboxを必ず返す
-5. 推測で補完せず、見えない情報は unknown を返す
-
-出力はJSON形式:
-{
-${SOURCE_LABEL_OUTPUT_SNIPPET}
-  "words": [
-    {
-      "english": "単語",
-      "japanese": "意味",
-      "partOfSpeechTags": ["noun"],
-      "markerColor": "yellow|pink|green|orange|blue|purple|red|black|pen|unknown",
-      "markType": "underline|highlight|unknown",
-      "isHandDrawn": true,
-      "confidence": 0.0〜1.0,
-      "wordBoundingBox": {"y_min": 0, "x_min": 0, "y_max": 1000, "x_max": 1000},
-      "markBoundingBox": {"y_min": 0, "x_min": 0, "y_max": 1000, "x_max": 1000}
-    }
-  ],
-  "detectedColors": ["green"],
-  "totalHighlightedRegions": 5
-}
-
-日本語訳が画像にあればそれを使って japaneseSource は "scan" を返してください。無ければ japanese は空文字 "" を返し、japaneseSource は付けないでください。推測生成はしないでください。
-partOfSpeechTags には noun / verb / adjective / adverb / idiom / phrasal_verb など、最も適切な分類を1つ以上入れてください。${SOURCE_LABEL_NOTES}`;
-
-export const HIGHLIGHTED_WORD_USER_PROMPT = `この画像で、手書きのマーカーやペンで印をつけた単語だけを抽出してください。印のない単語は含めないでください。sourceLabels には物理教材名だけを入れ、"鉄壁" や "LEAP" のような具体的書名だけを返してください。"英語教材" や "参考書" のような一般名詞は返さず、教材名不明のノート画像なら ["ノート"] を返してください。`;
-
-export const HIGHLIGHTED_WORD_VERIFICATION_SYSTEM_PROMPT = `あなたは抽出候補の再検証担当です。候補一覧にある単語だけを画像で再判定し、条件を満たすものだけ残してください。
-
-【残す条件】
-- 手書きマーカーまたは手書き下線が明確に確認できる
-- 下線の場合、線の真上にある単語である
-- 印刷赤字・印刷下線・書式装飾ではない
-
-【除外条件】
-- 下線の下の行の単語
-- 候補にない新規単語
-- 印が曖昧な候補
-
-出力はJSONのみ:
-{
-  "words": [
-    { "english": "word", "japanese": "意味", "japaneseSource": "scan" }
-  ]
-}`;
-
-// ============ Wrong Answer Extraction Prompts ============
-// For extracting only incorrectly answered words from vocabulary tests
-// Uses two-stage processing: Gemini for OCR → GPT for analysis
-
-export const WRONG_ANSWER_OCR_SYSTEM_PROMPT = `あなたは単語テストの画像を解析する専門家です。ユーザーがアップロードした単語テストの画像から、以下の情報を正確に抽出してください。
-
-═══════════════════════════════════════════════════════════════
-██  タスク: 単語テストの構造を完全に理解して抽出  ██
-═══════════════════════════════════════════════════════════════
-
-【単語テストの一般的な構造】
-単語テストには通常以下の要素があります：
-1. **問題番号**: 1, 2, 3... や (1), (2), (3)... など
-2. **問題（英単語または日本語）**: テストで問われている単語
-3. **解答欄**: 生徒が手書きで書いた答え
-4. **正解（あれば）**: 赤ペンで書かれた正解、または印刷された正解
-
-【重要な構造パターン】
-パターンA: 英→日テスト
-- 問題: 英単語が印刷されている
-- 解答欄: 日本語の意味を手書きで記入
-
-パターンB: 日→英テスト
-- 問題: 日本語の意味が印刷されている
-- 解答欄: 英単語を手書きで記入
-
-パターンC: 選択式テスト
-- 問題: 文や文脈が印刷されている
-- 選択肢: 複数の選択肢がある
-- 解答欄: 選択した番号や記号
-
-【採点マークの識別】
-- ○（丸）: 正解を示す
-- ×（バツ）: 不正解を示す
-- △（三角）: 部分点・惜しいを示す
-- 赤ペンでの修正: 正しい答えが書き加えられている
-- 点数: 「-1」「0点」など
-
-═══════════════════════════════════════════════════════════════
-██  出力フォーマット（JSON）  ██
-═══════════════════════════════════════════════════════════════
-
-{
-${SOURCE_LABEL_OUTPUT_SNIPPET}
-  "testType": "english_to_japanese | japanese_to_english | multiple_choice | mixed",
-  "questions": [
-    {
-      "questionNumber": 1,
-      "question": "問題文（英単語または日本語）",
-      "studentAnswer": "生徒が書いた答え（読み取れない場合は null）",
-      "correctAnswer": "正解（画像に記載されていれば。なければ null）",
-      "isCorrect": true | false | null,
-      "markingSymbol": "○ | × | △ | none | unclear",
-      "confidence": 0.0〜1.0
-    }
-  ],
-  "totalQuestions": 10,
-  "detectedCorrectCount": 7,
-  "detectedWrongCount": 3,
-  "notes": "特記事項（読み取りづらい部分など）"
-}
-
-【注意事項】
-- 手書き文字は完璧に読み取れないこともある。確信度(confidence)で示す
-- 採点マーク（○×△）を必ず探す
-- 赤ペンで書かれた修正・正解を見逃さない
-- 問題と解答欄の位置関係を正しく対応させる
-- 画像が不鮮明な部分は confidence を下げて出力
-${SOURCE_LABEL_RULES}
-${SOURCE_LABEL_NOTES}`;
-
-export const WRONG_ANSWER_OCR_USER_PROMPT = `この単語テストの画像を解析してください。
-
-解析手順：
-1. まずテスト全体の構造を把握する（英→日、日→英、選択式など）
-2. 各問題の番号、問題文、解答欄を特定する
-3. 採点マーク（○×△）や赤ペンの修正を探す
-4. 生徒の手書き解答を読み取る
-5. 正解が書かれていれば、それも抽出する
-
-重要：問題文と解答欄が離れた位置にある場合も、問題番号を頼りに正しく対応させてください。`;
-
-export const WRONG_ANSWER_ANALYSIS_SYSTEM_PROMPT = `あなたは英語学習教材の作成者です。OCRで抽出された単語テストの結果から、**間違えた単語のみ**を特定し、単語帳に登録するためのデータを生成してください。
-
-═══════════════════════════════════════════════════════════════
-██  最重要命令：間違いだけを抽出せよ  ██
-═══════════════════════════════════════════════════════════════
-
-【抽出対象】
-✓ ×（バツ）マークがついた問題
-✓ △（三角）マークがついた問題（部分的な間違い）
-✓ 赤ペンで修正が入っている問題
-✓ isCorrect が false の問題
-
-【抽出対象外】
-✗ ○（丸）マークがついた正解の問題
-✗ isCorrect が true の問題
-✗ 採点されていないが、正しいと判断できる問題
-
-═══════════════════════════════════════════════════════════════
-██  単語データの生成ルール  ██
-═══════════════════════════════════════════════════════════════
-
-【英単語と日本語訳の決定】
-1. テストタイプが「english_to_japanese」の場合:
-   - english: 問題文（テストに印刷されている英単語）
-   - japanese: 正解として示されている日本語訳。画像に日本語の正解が見えない場合は空文字 "" にする
-
-2. テストタイプが「japanese_to_english」の場合:
-   - english: 正解として示されている英単語、またはAIが生成
-   - japanese: 問題文（テストに印刷されている日本語）
-
-3. 日本語訳が画像に含まれている場合: そのまま使用し、japaneseSource は "scan" を返す
-4. english_to_japanese で日本語訳が画像にない場合: japanese は空文字 "" にし、japaneseSource は付けない
-5. japanese_to_english で英単語の正解が画像にない場合のみ、english をAIで補完してよい
-6. 各語には最も適切な主分類を1つだけ partOfSpeechTags に入れる
-
-【禁止事項】
-- 正解した問題を出力しない
-
-═══════════════════════════════════════════════════════════════
-██  出力フォーマット（JSON）  ██
-═══════════════════════════════════════════════════════════════
-
-{
-  "sourceLabels": ["鉄壁"],
-  "words": [
-    {
-      "english": "英単語",
-      "japanese": "日本語訳",
-      "japaneseSource": "scan",
-      "partOfSpeechTags": ["noun"],
-      "studentMistake": "生徒が間違えて書いた答え（参考用）",
-      "questionNumber": 3
-    }
-  ],
-  "summary": {
-    "totalWrong": 3,
-    "testType": "english_to_japanese",
-    "suggestions": "この生徒へのアドバイス（任意）"
-  }
-}
-
-【注意事項】
-- 間違えた問題が見つからない場合は {"words": [], "summary": {"totalWrong": 0}} を返す
-- 採点が不明確な場合は、安全のため「間違い」として扱う（復習して損はない）${SOURCE_LABEL_NOTES}`;
-
-export const WRONG_ANSWER_ANALYSIS_USER_PROMPT = `以下のOCR結果から、間違えた単語のみを抽出し、単語帳に登録するためのデータを生成してください。
-
-OCR結果:
-`;
+export {
+  WORD_EXTRACTION_SYSTEM_PROMPT,
+  USER_PROMPT_TEMPLATE,
+  WORD_EXTRACTION_WITH_EXAMPLES_SYSTEM_PROMPT,
+  USER_PROMPT_WITH_EXAMPLES_TEMPLATE,
+} from './prompts/word-extraction';
+
+export {
+  CIRCLED_WORD_EXTRACTION_SYSTEM_PROMPT,
+  CIRCLED_WORD_USER_PROMPT,
+  CIRCLED_WORD_VERIFICATION_SYSTEM_PROMPT,
+} from './prompts/circled';
+
+export {
+  EIKEN_LEVEL_DESCRIPTIONS,
+  getEikenLevelsAbove,
+  getEikenFilterInstruction,
+  EIKEN_OCR_PROMPT,
+  EIKEN_WORD_ANALYSIS_SYSTEM_PROMPT,
+  EIKEN_WORD_ANALYSIS_USER_PROMPT,
+} from './prompts/eiken';
+
+export {
+  GRAMMAR_OCR_PROMPT,
+  GRAMMAR_ANALYSIS_SYSTEM_PROMPT,
+  GRAMMAR_ANALYSIS_USER_PROMPT,
+  getGrammarLevelFilterInstruction,
+} from './prompts/grammar';
+
+export {
+  IDIOM_EXTRACTION_SYSTEM_PROMPT,
+  IDIOM_USER_PROMPT,
+} from './prompts/idiom';
+
+export {
+  HIGHLIGHTED_WORD_EXTRACTION_SYSTEM_PROMPT,
+  HIGHLIGHTED_WORD_USER_PROMPT,
+  HIGHLIGHTED_WORD_VERIFICATION_SYSTEM_PROMPT,
+} from './prompts/highlighted';
+
+export {
+  WRONG_ANSWER_OCR_SYSTEM_PROMPT,
+  WRONG_ANSWER_OCR_USER_PROMPT,
+  WRONG_ANSWER_ANALYSIS_SYSTEM_PROMPT,
+  WRONG_ANSWER_ANALYSIS_USER_PROMPT,
+} from './prompts/wrong-answer';
