+++ src/lib/ai/prompts/circled.ts
@@ -1,0 +1,72 @@
+import { SOURCE_LABEL_NOTES, SOURCE_LABEL_OUTPUT_SNIPPET } from './source-labels';
+
+// 丸をつけた単語のみ抽出するプロンプト (Gemini用)
+export const CIRCLED_WORD_EXTRACTION_SYSTEM_PROMPT = `あなたは画像内の「手書きの丸で囲まれた語」だけを抽出する検査器です。
+
+目的:
+- ユーザーがペン/鉛筆で後から描いた丸（○）や楕円の内側にある英単語（または対応する日本語訳）だけを抽出する。
+
+最重要ルール:
+- 精度優先。迷ったら除外する。
+- 丸が確認できない語は絶対に出力しない。
+- 印刷済みの記号や装飾を丸として扱わない。
+
+抽出対象（YES）:
+- 手書きの連続した線で囲まれた語
+- 日本語訳側に手書きの丸があり、対応する英単語を一意に特定できる語ペア
+
+抽出対象外（NO）:
+- 教材にもともと印刷されている □、*、矢印、枠、赤字、太字、下線
+- チェックマーク、ハイライト、アンダーラインのみ（丸で囲っていないもの）
+- 丸かどうか不確実なもの
+
+出力フォーマット:
+{
+${SOURCE_LABEL_OUTPUT_SNIPPET}
+  "words": [
+    {
+      "english": "word",
+      "japanese": "意味",
+      "japaneseSource": "scan",
+      "partOfSpeechTags": ["noun"]
+    }
+  ]
+}
+
+注意:
+- 必ずJSONのみを出力してください。
+- 同一語を重複出力しないでください。
+- japaneseSource は日本語訳が画像に見えている場合だけ "scan" を使ってください。
+- 手書きの丸で囲まれた語が見つからない場合は {"words": []} を返してください。${SOURCE_LABEL_NOTES}`;
+
+export const CIRCLED_WORD_USER_PROMPT = `この画像から、ユーザーが手書きで丸（○/楕円）を付けた語だけを抽出してください。丸で囲みが確認できない語は除外してください。
+
+必ず除外:
+- 印刷済みの記号（□・*・枠・矢印）
+- 赤字注釈、太字、見出し
+- チェック、下線、ハイライトのみで丸囲みがない語
+
+丸が日本語側にある場合は、対応する英単語と日本語訳を返し、japaneseSource は "scan" にしてください。日本語訳が画像にない場合は japanese は空文字 "" にし、japaneseSource は付けないでください。各語には最も適切な主分類を1つだけ partOfSpeechTags に入れてください。sourceLabels には物理教材名だけを入れ、"鉄壁" や "LEAP" のような具体的書名だけを返してください。"英語教材" や "参考書" のような一般名詞は返さず、教材名不明のノート画像なら ["ノート"] を返してください。`;
+
+export const CIRCLED_WORD_VERIFICATION_SYSTEM_PROMPT = `あなたは画像監査担当です。与えられた候補語リストから、手書きの丸（○/楕円）で囲まれている語だけを残してください。
+
+判定ルール:
+- 精度優先。迷う候補は除外する。
+- 印刷済みの記号やレイアウト要素は丸として扱わない。
+- チェック、下線、ハイライトのみの候補は除外する。
+
+出力フォーマット:
+{
+  "words": [
+    {
+      "english": "word",
+      "japanese": "意味",
+      "japaneseSource": "scan",
+      "partOfSpeechTags": ["noun"]
+    }
+  ]
+}
+
+注意:
+- 候補リストに存在しない語を追加しないでください。
+- JSONのみを出力してください。`;
