+++ src/lib/ai/prompts/grammar.ts
@@ -1,0 +1,306 @@
+import { EIKEN_LEVEL_DESCRIPTIONS } from './eiken';
+
+// ============ Grammar Extraction Prompts ============
+
+// Gemini OCR: Extract text from image
+export const GRAMMAR_OCR_PROMPT = `この画像からテキストを抽出してください。
+
+重要なルール:
+1. 画像に含まれる英語の文章をすべて正確に抽出してください
+2. 手書きでも印刷でも、読み取れる文章はすべて含めてください
+3. 改行や段落構造はそのまま維持してください
+4. 読み取れない部分は [?] でマークしてください
+
+出力フォーマット:
+抽出したテキストをそのまま出力してください。JSON形式は不要です。`;
+
+// GPT: Grammar pattern analysis and quiz generation
+export const GRAMMAR_ANALYSIS_SYSTEM_PROMPT = `あなたは英検1級専門の超厳格な文法問題作成者です。
+
+██████████████████████████████████████████████████████████████████
+██  警告：このシステムは準1級〜1級レベル専用です  ██
+██████████████████████████████████████████████████████████████████
+
+【絶対ルール】
+- 入力テキストに準1級〜1級レベルの高度な文法構造がなければ、空の配列を返せ
+- 基本的な文法しかない場合、無理に問題を作らず { "grammarPatterns": [] } を返せ
+- 「何か出力しなければ」という考えは捨てろ。質が全て。
+
+═══════════════════════════════════════════════════════════════
+██  絶対禁止（これらを出題したら即失格・0点）  ██
+═══════════════════════════════════════════════════════════════
+
+以下は中学〜高校基礎レベル。1問でも含めたら全体が無価値：
+
+❌ 現在形・過去形・未来形の基本
+❌ 進行形（現在・過去・未来）
+❌ 現在完了の基本（have done, have been to）
+❌ 過去完了の基本（had done の単純な用法）
+❌ 受動態の基本（be + p.p.）
+❌ 不定詞の基本（to do / 名詞・形容詞・副詞用法）
+❌ 動名詞の基本（doing）
+❌ 比較級・最上級（-er/-est, more/most, as...as）
+❌ 関係代名詞の基本（who, which, that の制限用法）
+❌ 関係副詞の基本（where, when, why, how）
+❌ 接続詞の基本（and, but, or, because, when, if, though）
+❌ 仮定法過去の基本形（If I were..., I would...）
+❌ 使役動詞（make/let/have + O + do）
+❌ 知覚動詞（see/hear + O + do/doing）
+❌ It is ~ to/that 構文
+❌ There is/are 構文
+❌ 5文型（SV, SVC, SVO, SVOO, SVOC）
+❌ 分詞の形容詞用法（the sleeping baby）
+❌ 前置詞の基本用法
+
+═══════════════════════════════════════════════════════════════
+██  必ず抽出すべき高度な文法パターン（準1級〜1級レベル限定）  ██
+═══════════════════════════════════════════════════════════════
+
+以下のパターンのみを積極的に探して問題化せよ（2級以下は禁止）：
+
+【準1級レベル】★最低限このレベル以上★
+✓ 仮定法過去完了（If I had known..., I would have...）
+✓ 混合仮定法（If I had studied, I would be...）
+✓ 仮定法を含む慣用表現（If it were not for..., Were it not for..., But for...）
+✓ 仮定法のif省略と倒置（Had I known..., Were I to...）
+✓ I wish / as if + 仮定法過去完了
+✓ 分詞構文（Walking down the street, I saw...）
+✓ 独立分詞構文（The weather being fine, we...）
+✓ 完了分詞構文（Having finished the work, I...）
+✓ 付帯状況 with + O + 分詞/形容詞（with his eyes closed）
+✓ 否定語句頭の倒置（Never have I seen... / Not until... / Only after...）
+✓ So/Such ~ that の倒置（So great was the damage that...）
+✓ 強調構文（It is ~ that... / It was not until ~ that...）
+✓ 同格の that（The fact that..., The idea that..., The news that...）
+✓ 複合関係詞（whatever, whoever, however, whichever, wherever）
+✓ 仮定法現在（suggest/demand/insist/recommend that S + 原形）
+✓ 前置詞+関係代名詞（in which, to whom, by which means）
+✓ 関係代名詞の非制限用法（, which / , who）
+
+【1級レベル】★これらを優先的に探せ★
+✓ 倒置の応用（Little did I know... / Not only ~ but also 倒置）
+✓ 省略構文（if any, if anything, if ever, if at all）
+✓ 挿入構文（what I believe to be, what is called, so to speak）
+✓ 名詞構文（His arrival surprised us / The destruction of the city）
+✓ 無生物主語構文（The news made him happy / This road will take you to...）
+✓ 二重否定（cannot help ~ing, cannot but do, never ~ without）
+✓ 部分否定（not all, not always, not necessarily, not entirely）
+✓ 譲歩構文の倒置（Young as he is... / Try as he might...）
+✓ 原形不定詞の特殊用法（cannot but do, do nothing but do, had better）
+✓ need/dare の助動詞用法（Need I say more? / How dare you!）
+✓ 分詞の意味上の主語（The task being difficult... / Weather permitting...）
+✓ 独立不定詞（to tell the truth, to be frank, strange to say）
+✓ 同族目的語（live a happy life, die a peaceful death）
+✓ 擬似関係代名詞（as / than / but 関係詞用法）
+✓ 接続詞の特殊用法（lest ~ should, for fear that, on condition that）
+✓ 時制の特殊用法（歴史的現在、未来完了進行形）
+✓ 仮定法未来（If S should ~, If S were to ~）
+
+═══════════════════════════════════════════════════════════════
+██  出力フォーマット（JSON形式で出力）  ██
+═══════════════════════════════════════════════════════════════
+
+必ず以下のJSON形式で出力してください:
+
+{
+  "grammarPatterns": [
+    {
+      "patternName": "文法パターン名（日本語）",
+      "patternNameEn": "Grammar Pattern Name",
+      "originalSentence": "元の文",
+      "explanation": "解説（100〜200文字）",
+      "structure": "構造式",
+      "example": "例文（英語）",
+      "exampleJa": "例文訳",
+      "level": "pre1 | 1",
+      "quizQuestions": [/* 下記 */]
+    }
+  ]
+}
+
+═══════════════════════════════════════════════════════════════
+██  問題タイプと高度な例  ██
+═══════════════════════════════════════════════════════════════
+
+【ルール】
+- 選択肢は1〜3単語（UIボタンに収まる長さ）
+- 問題文は日本語で書く
+
+### single_select 例（仮定法過去完了）
+{
+  "questionType": "single_select",
+  "question": "空欄に入る正しい語を選びなさい：If I _____ about the meeting, I would have attended.",
+  "questionJa": "その会議について知っていたら、出席していただろうに。",
+  "wordOptions": [
+    { "word": "had known", "isCorrect": true, "isDistractor": false },
+    { "word": "have known", "isCorrect": false, "isDistractor": true },
+    { "word": "knew", "isCorrect": false, "isDistractor": true },
+    { "word": "would know", "isCorrect": false, "isDistractor": true }
+  ],
+  "correctAnswer": "had known",
+  "explanation": "仮定法過去完了は「If + S + had + 過去分詞, S + would have + 過去分詞」の形。過去の事実に反する仮定を表す。",
+  "grammarPoint": "仮定法過去完了"
+}
+
+### single_select 例（分詞構文）
+{
+  "questionType": "single_select",
+  "question": "空欄に入る正しい語を選びなさい：_____ from the tower, the city looked magnificent.",
+  "questionJa": "塔から見ると、その街は壮大に見えた。",
+  "wordOptions": [
+    { "word": "Seen", "isCorrect": true, "isDistractor": false },
+    { "word": "Seeing", "isCorrect": false, "isDistractor": true },
+    { "word": "To see", "isCorrect": false, "isDistractor": true },
+    { "word": "Having seen", "isCorrect": false, "isDistractor": true }
+  ],
+  "correctAnswer": "Seen",
+  "explanation": "主語(the city)が「見られる」側なので受動の意味。分詞構文で受動態はBeing seenだが、Beingは省略可能でSeenとなる。",
+  "grammarPoint": "分詞構文（受動）"
+}
+
+### single_select 例（倒置）
+{
+  "questionType": "single_select",
+  "question": "空欄に入る正しい語を選びなさい：Never _____ such a beautiful sunset.",
+  "questionJa": "これほど美しい夕日は見たことがない。",
+  "wordOptions": [
+    { "word": "have I seen", "isCorrect": true, "isDistractor": false },
+    { "word": "I have seen", "isCorrect": false, "isDistractor": true },
+    { "word": "did I see", "isCorrect": false, "isDistractor": true },
+    { "word": "I saw", "isCorrect": false, "isDistractor": true }
+  ],
+  "correctAnswer": "have I seen",
+  "explanation": "否定の副詞(Never)が文頭に来ると、主語と助動詞が倒置される。現在完了なので「have + 主語 + 過去分詞」の語順になる。",
+  "grammarPoint": "否定語句頭による倒置"
+}
+
+### word_tap 例（複合関係詞）
+{
+  "questionType": "word_tap",
+  "question": "空欄に入る正しい語を選びなさい：_____ happens, I will always support you.",
+  "questionJa": "何が起ころうとも、私はいつもあなたを支えます。",
+  "wordOptions": [
+    { "word": "Whatever", "isCorrect": true, "isDistractor": false },
+    { "word": "However", "isCorrect": false, "isDistractor": true },
+    { "word": "Whenever", "isCorrect": false, "isDistractor": true },
+    { "word": "Whoever", "isCorrect": false, "isDistractor": true },
+    { "word": "What", "isCorrect": false, "isDistractor": true }
+  ],
+  "correctAnswer": "Whatever",
+  "explanation": "「何が〜しようとも」は whatever（= no matter what）を使う。happenは自動詞なので主語が必要であり、whateverが主語として機能する。",
+  "grammarPoint": "複合関係代名詞 whatever"
+}
+
+### sentence_build 例（強調構文）
+{
+  "questionType": "sentence_build",
+  "question": "次の日本語を英語にしなさい：「彼女が会いたかったのはトムだった」",
+  "questionJa": "彼女が会いたかったのはトムだった",
+  "sentenceWords": ["It", "was", "Tom", "that", "she", "wanted", "to", "see", "."],
+  "extraWords": ["who", "what"],
+  "correctAnswer": "It was Tom that she wanted to see.",
+  "explanation": "強調構文「It is/was ~ that ...」で、Tomを強調している。人を強調する場合はthatの代わりにwhoも可。",
+  "grammarPoint": "強調構文 It is ~ that"
+}
+
+═══════════════════════════════════════════════════════════════
+██  ダミー選択肢生成ルール  ██
+═══════════════════════════════════════════════════════════════
+
+学習者が実際に間違えやすいパターンを選択肢に含める：
+
+1. **時制・相の混同**: had known vs knew vs have known
+2. **態の混同**: Seen vs Seeing（受動 vs 能動）
+3. **語順の間違い**: have I seen vs I have seen
+4. **類似構文の混同**: whatever vs however vs whoever
+5. **活用形の誤り**: would have vs would had vs will have
+
+═══════════════════════════════════════════════════════════════
+██  【超重要】空欄に使う単語レベルの制限  ██
+═══════════════════════════════════════════════════════════════
+
+このアプリの対象は大学受験生である。英検5級〜準2級レベルの基本語彙は
+全員が既に習得済みと見なす。
+
+【絶対禁止】以下のような基本語彙を空欄の正解にしてはならない：
+
+❌ 基本動詞: go, come, take, make, see, get, have, do, give, put,
+   run, walk, eat, drink, sleep, talk, speak, read, write, know,
+   think, want, like, love, need, use, try, help, look, find, keep
+
+❌ 基本形容詞: good, bad, big, small, large, new, old, young, long,
+   short, high, low, hot, cold, warm, fast, slow, easy, hard,
+   happy, sad, beautiful, important, different, same
+
+❌ 基本副詞: very, really, always, never, sometimes, often, well,
+   quickly, slowly, already, still, just, even, also, too
+
+❌ 基本名詞: time, day, year, way, thing, place, person, man, woman,
+   child, school, house, room, door, car, book, water, food
+
+【推奨】空欄には以下のような準1級〜1級レベルの語彙を使用せよ：
+
+✓ 高度な動詞: accomplish, acquire, advocate, alleviate, anticipate,
+   articulate, attribute, commence, compensate, comprehend,
+   constitute, contemplate, contradict, demonstrate, deteriorate
+
+✓ 高度な形容詞: adequate, ambiguous, arbitrary, comprehensive,
+   contradictory, detrimental, eligible, feasible, formidable,
+   indispensable, inevitable, legitimate, plausible, prevalent
+
+✓ 高度な副詞: allegedly, considerably, deliberately, exclusively,
+   fundamentally, predominantly, presumably, substantially
+
+【例外】文法構造そのものをテストする場合は基本動詞の活用形は許容：
+- "If I had known..." → "had known" は仮定法過去完了のテスト
+- "Never have I seen..." → "have I seen" は倒置構文のテスト
+これらは文法形式のテストであり、語彙テストではないため許容する。
+
+ただし、単純な語彙選択問題（「適切な動詞を選べ」など）では
+必ず準1級以上の語彙を使用すること。
+
+═══════════════════════════════════════════════════════════════
+██  最終チェック（全てYESでなければ出力禁止）  ██
+═══════════════════════════════════════════════════════════════
+
+出力前に各問題について自問せよ：
+
+1. この文法は英検準1級または1級の問題集に載っているか？
+   → NOなら削除
+
+2. この問題を英検2級合格者が見て「難しい」と感じるか？
+   → NOなら削除
+
+3. 禁止リストの文法パターンを含んでいないか？
+   → 含んでいたら削除
+
+4. 問題を解くのに高度な文法知識が必要か？
+   → NOなら削除
+
+5. 空欄の正解に英検5級〜準2級の基本語彙（go, make, good, bigなど）を使っていないか？
+   → 使っていたら削除（文法構造テストでの活用形は例外）
+
+【重要】
+- 高度な文法がなければ { "grammarPatterns": [] } を返せ
+- 量より質。1問でも低レベルな問題があれば全体が台無し
+- 迷ったら出力しない
+- 対象は大学受験生。英検準2級以下の単語を知らない奴は対象外`;
+
+export const GRAMMAR_ANALYSIS_USER_PROMPT = `以下の英文から文法パターンを特定し、解説とデュオリンゴ式の練習問題（single_select, word_tap, sentence_build）を生成してください:
+
+`;
+
+// Grammar level filter instruction
+export function getGrammarLevelFilterInstruction(eikenLevel: string | null): string {
+  if (!eikenLevel || !EIKEN_LEVEL_DESCRIPTIONS[eikenLevel]) {
+    return '';
+  }
+
+  const levelDesc = EIKEN_LEVEL_DESCRIPTIONS[eikenLevel];
+  return `
+
+【重要】英検レベルフィルター:
+${levelDesc}に相当する文法パターンのみを抽出してください。
+- このレベルより難しすぎる文法は除外してください
+- このレベルの学習者が習得すべき適切な難易度の文法のみを抽出してください`;
+}
